Directives can include dots and dashes
A directive name can include dots (.) and dashes (-).  The parser wasn't
handling this.  It is unclear if a directive in an assignment can have
other characters.  The docs talk about the directive "consists of the
name and parameters separated by spaces and ends with a semicolon (;)."
(http://nginx.org/en/docs/beginners_guide.html).  I'm not sure if this
means everything up to the next space is legal.